MB-67250: Fix narrowing warnings in ep-engine_ep_unit_tests
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230969
Tested-by: Jim Walker <jim@couchbase.com>

+++ engines/ep/tests/mock/mock_stream.h
@@ -121,7 +121,7 @@ public:
         transitionState(StreamState::Dead);
     }
 
-    int getNumBackfillItems() const {
+    size_t getNumBackfillItems() const {
         return backfillItems.memory + backfillItems.disk;
     }
