+++ engines/ep/tests/module_tests/nexus_kvstore_test.cc
@@ -545,7 +545,7 @@ TEST_P(NexusKVStoreTest, SecondaryExpiresFromSameTime) {
 
     // Expire this item in the future
     auto expiredKey = makeStoredDocKey("key");
-    auto timeNow = ep_real_time();
+    auto timeNow = gsl::narrow<uint32_t>(ep_real_time());
     store_item(vbid, expiredKey, "value", timeNow + 10 /*exptime*/);
     flushVBucketToDiskIfPersistent(vbid, 1);
